IDK how to read stream reader

+++ MusicGradeApp/MusicGradeApp/IGenre.cs
@@ -10,7 +10,7 @@ namespace MusicGradeApp
     {
         string MusicGenre { get; }
 
-        void AddTrack(string title, float rating);
+        void AddTrack(string title, int rating);
         void AddTrack(string title, string rating);
         Statistics GetStatistics();
     }
